move files around make helper function for ValidatePasswords make new LoginForm component
